+++ src/bika/lims/utils/analysis.py
@@ -18,6 +18,7 @@
 # Copyright 2018-2021 by it's authors.
 # Some rights reserved, see README and LICENSE.
 
+import copy
 import math
 
 from bika.lims import api
@@ -26,6 +27,7 @@ from bika.lims.interfaces import IBaseAnalysis
 from bika.lims.interfaces import IReferenceSample
 from bika.lims.interfaces.analysis import IRequestAnalysis
 from bika.lims.utils import formatDecimalMark
+from bika.lims.utils import format_supsub
 
 
 def create_analysis(context, source, **kwargs):
@@ -416,3 +418,42 @@ def generate_analysis_id(instance, keyword):
         new_id = "{}-{}".format(keyword, count)
         count += 1
     return new_id
+
+
+def format_interim(interim_field, html=True):
+    """Returns a copy of the interim field plus additional attributes suitable
+    for visualization, like formatted_result and formatted_unit
+    """
+    separator = "<br/>" if html else ", "
+
+    # copy to prevent persistent changes
+    item = copy.deepcopy(interim_field)
+
+    # get the raw value
+    value = item.get("value", "")
+    values = filter(None, api.to_list(value))
+
+    # if choices, display texts instead of values
+    choices = item.get("choices")
+    if choices:
+        # generate a {value:text} dict
+        choices = choices.split("|")
+        choices = dict(map(lambda ch: ch.strip().split(":"), choices))
+
+        # set the text as the formatted value
+        texts = [choices.get(v, "") for v in values]
+        values = filter(None, texts)
+
+    else:
+        # default formatting
+        setup = api.get_setup()
+        decimal_mark = setup.getResultsDecimalMark()
+        values = [formatDecimalMark(val, decimal_mark) for val in values]
+
+    item["formatted_value"] = separator.join(values)
+
+    # unit formatting
+    unit = item.get("unit", "")
+    item["formatted_unit"] = format_supsub(unit) if html else unit
+
+    return item